Add files via upload
This project analyzed real estate in Russian cities. Various parameters were used in the analysis, such as location, number of rooms in the apartment, area and height of rooms, distance from public places, etc. By comparing real estate in different cities, taking into account all possible factors, we have proposed options where it is best to buy an apartment in order to quickly resell it in the future.